[#192] 구인구직 테스트코드 추가 (#193)
* [#192] 구인구직 테스트코드 기반 및 등록 API 추가

* [#192] user common call api 네이밍 변경

* [#192] 구인구직 삭제 테스트코드 추가

* [#192] lint 적용

* [#192] 임시저장

* [#192] 구인구직 수정 테스트코드 추가

* [#192] 구인구직 조회 테스트코드 추가

* [#192] 내가 등록한 구인구직 조회 테스트코드 추가

* [#99] java.lang.IllegalStateException: Session/EntityManager is closed 에러 수정 (#197)

* [#192] 필요없는 변수 정리

* [#192] 구인구직 스크랩 API 테스트 코드 추가

* [#192] 구인구직 스크랩 조회 API 테스트 코드 추가

* [#192] 나와 비슷한 사람들이 보고 있는 공고 조회 API 테스트 코드 추가

* [#192] 테스트코드 오류 수정

* [#192] 삭제된 구인구직 조회안되게 변경

+++ server/src/test/kotlin/com/fone/common/WebTestClientExtension.kt
@@ -16,7 +16,10 @@ fun WebTestClient.doGet(
         return this.get().uri { it.setUriBuilder(url, queryParams) }.exchange()
     }
 
-    return this.get().uri { it.setUriBuilder(url, queryParams) }.headers { it.setBearerAuth(token) }.exchange()
+    return this.get()
+        .uri { it.setUriBuilder(url, queryParams) }
+        .headers { it.setBearerAuth(token) }
+        .exchange()
 }
 
 fun <T> WebTestClient.doPost(
@@ -26,26 +29,49 @@ fun <T> WebTestClient.doPost(
     queryParams: Map<String, Any>? = null,
 ): WebTestClient.ResponseSpec {
     if (request == null && token != null) {
-        return this.post().uri { it.setUriBuilder(url, queryParams) }.contentType(MediaType.APPLICATION_JSON)
-            .headers { it.setBearerAuth(token) }.exchange()
+        return this.post()
+            .uri { it.setUriBuilder(url, queryParams) }
+            .contentType(MediaType.APPLICATION_JSON)
+            .headers { it.setBearerAuth(token) }
+            .exchange()
     }
 
     if (token == null) {
-        return this.post().uri { it.setUriBuilder(url, queryParams) }.contentType(MediaType.APPLICATION_JSON)
-            .body(BodyInserters.fromValue(request)).exchange()
+        return this.post()
+            .uri { it.setUriBuilder(url, queryParams) }
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(BodyInserters.fromValue(request))
+            .exchange()
     }
 
-    return this.post().uri { it.setUriBuilder(url, queryParams) }.contentType(MediaType.APPLICATION_JSON)
-        .body(BodyInserters.fromValue(request)).headers { it.setBearerAuth(token) }.exchange()
+    return this.post()
+        .uri { it.setUriBuilder(url, queryParams) }
+        .contentType(MediaType.APPLICATION_JSON)
+        .body(BodyInserters.fromValue(request))
+        .headers { it.setBearerAuth(token) }
+        .exchange()
 }
 
 fun <T> WebTestClient.doPut(
     url: String,
     request: T,
+    token: String? = null,
     queryParams: Map<String, Any>? = null,
 ): WebTestClient.ResponseSpec {
-    return this.put().uri { it.setUriBuilder(url, queryParams) }.contentType(MediaType.APPLICATION_JSON)
-        .body(BodyInserters.fromValue(request)).exchange()
+    if (token == null) {
+        return this.put()
+            .uri { it.setUriBuilder(url, queryParams) }
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(BodyInserters.fromValue(request))
+            .exchange()
+    }
+
+    return this.put()
+        .uri { it.setUriBuilder(url, queryParams) }
+        .contentType(MediaType.APPLICATION_JSON)
+        .body(BodyInserters.fromValue(request))
+        .headers { it.setBearerAuth(token) }
+        .exchange()
 }
 
 fun <T> WebTestClient.doPatch(
@@ -55,35 +81,55 @@ fun <T> WebTestClient.doPatch(
     queryParams: Map<String, Any>? = null,
 ): WebTestClient.ResponseSpec {
     if (request == null && token != null) {
-        return this.patch().uri { it.setUriBuilder(url, queryParams) }.contentType(MediaType.APPLICATION_JSON)
-            .headers { it.setBearerAuth(token) }.exchange()
+        return this.patch()
+            .uri { it.setUriBuilder(url, queryParams) }
+            .contentType(MediaType.APPLICATION_JSON)
+            .headers { it.setBearerAuth(token) }
+            .exchange()
     }
 
     if (token == null) {
-        return this.patch().uri { it.setUriBuilder(url, queryParams) }.contentType(MediaType.APPLICATION_JSON)
-            .body(BodyInserters.fromValue(request)).exchange()
+        return this.patch()
+            .uri { it.setUriBuilder(url, queryParams) }
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(BodyInserters.fromValue(request))
+            .exchange()
     }
 
-    return this.patch().uri { it.setUriBuilder(url, queryParams) }.contentType(MediaType.APPLICATION_JSON)
-        .body(BodyInserters.fromValue(request)).headers { it.setBearerAuth(token) }.exchange()
+    return this.patch()
+        .uri { it.setUriBuilder(url, queryParams) }
+        .contentType(MediaType.APPLICATION_JSON)
+        .body(BodyInserters.fromValue(request))
+        .headers { it.setBearerAuth(token) }
+        .exchange()
 }
 
 fun WebTestClient.doDelete(
     url: String,
+    token: String? = null,
     queryParams: Map<String, Any>? = null,
 ): WebTestClient.ResponseSpec {
-    return this.delete().uri { it.setUriBuilder(url, queryParams) }.exchange()
+    if (token == null) {
+        return this.delete().uri { it.setUriBuilder(url, queryParams) }.exchange()
+    }
+
+    return this.delete()
+        .uri { it.setUriBuilder(url, queryParams) }
+        .headers { it.setBearerAuth(token) }
+        .exchange()
 }
 
 private fun UriBuilder.setUriBuilder(url: String, queryParams: Map<String, Any>? = null): URI {
-    return this.path(url).run {
-        if (queryParams != null) {
-            this.queryParams(
-                LinkedMultiValueMap<String, String>().apply {
-                    setAll(queryParams.mapValues { it.value.toString() })
-                }
-            )
+    return this.path(url)
+        .run {
+            if (queryParams != null) {
+                this.queryParams(
+                    LinkedMultiValueMap<String, String>().apply {
+                        setAll(queryParams.mapValues { it.value.toString() })
+                    }
+                )
+            }
+            this
         }
-        this
-    }.build()
+        .build()
 }
